requirements.txt
# Web Framework
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.22

# Datenvalidierung
pydantic==2.10.4
pydantic-settings==2.7.1
pydantic[email]

# Datenbank
sqlalchemy==2.0.36
aiosqlite==0.20.0

# HTTP Client (für Ollama, paperless)
httpx==0.28.1

# PDF-Verarbeitung (falls benötigt)
pypdf==5.1.0

# Utilities
python-dotenv==1.0.1
pyyaml==6.0.2
jinja2==3.1.5

# Development
pytest==8.3.4
pytest-asyncio==0.25.2
